Amelioration de l'api spring boot

+++ src/main/java/com/ide/api/service/FileService.java
@@ -1,17 +1,12 @@
 package com.ide.api.service;
 
 import com.ide.api.configurations.FilePaths;
-import com.ide.api.configurations.JwtTokenProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.io.*;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
+import java.nio.file.*;
 import java.util.Base64;
 import java.util.Set;
 
@@ -34,15 +29,13 @@ public class FileService {
             byte[] decoder = Base64.getDecoder().decode(base64);
             fileOutputStream.write(decoder);
             fileOutputStream.flush();
-            fileOutputStream.close();
         }catch (Exception ex){
             ex.printStackTrace();
         }
     }
 
-    public String getFile(String nomFichier) throws NoSuchAlgorithmException, InvalidKeyException {
-        String base64Fichier = encoderFichierToBase64(nomFichier);
-        return base64Fichier;
+    public String getFile(String nomFichier){
+        return encoderFichierToBase64(nomFichier);
     }
 
     public byte[] getByteFile(String fileId) {
@@ -77,6 +70,25 @@ public class FileService {
     }
 
 
+    public boolean isH264Video(String filePath) throws IOException {
+        ProcessBuilder processBuilder = new ProcessBuilder(
+                "ffprobe",
+                "-v", "error",
+                "-select_streams", "v:0",
+                "-show_entries", "stream=codec_name",
+                "-of", "default=nw=1:nk=1",
+                filePath
+        );
+
+        Process process = processBuilder.start();
+        try (InputStream inputStream = process.getInputStream()) {
+            BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
+            String codecName = reader.readLine();
+            return "h264".equals(codecName);
+        }
+    }
+
+
 
     public void reencodeVideoToH264(String inputFilePath, String outputFilePath) throws IOException {
         logger.info("Entrer dans la fonction reencodage!");
@@ -84,21 +96,25 @@ public class FileService {
                 "ffmpeg",
                 "-i", inputFilePath,
                 "-vcodec", "libx264",
-                "-preset", "fast",  // Choisir un preset pour une meilleure performance
+                "-preset", "fast",
+                "-crf", "23",
+                "-vf", "scale='min(1280,iw)':-2",
+                "-tune", "film",
                 "-acodec", "aac",
-                "-b:a", "128k",     // Débit audio
-                "-strict", "experimental", // Si nécessaire pour utiliser certains codecs expérimentaux
+                "-b:a", "128k",
+                "-movflags", "+faststart",
+                "-strict", "experimental",
                 outputFilePath
         );
 
         // Rediriger les sorties de processus
-        processBuilder.redirectErrorStream(true); // Combinaison des flux stdout et stderr
+        processBuilder.redirectErrorStream(true);
         Process process = processBuilder.start();
         try (InputStream inputStream = process.getInputStream()) {
             BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
             String line;
             while ((line = reader.readLine()) != null) {
-                logger.info(line); // Logger chaque ligne de sortie
+                logger.info(line);
             }
         }
 
@@ -108,24 +124,24 @@ public class FileService {
             process.waitFor();
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
-            throw new RuntimeException("Interrupted while waiting reencode video", e);
         }
     }
 
     // Méthode pour stocker un fichier en fonction de son type (réencodage vidéo uniquement si vidéo)
     public void storeFileWithReencoding(String nomFichier, String dossierFichier, String base64) throws IOException {
-        // Si le fichier est une vidéo, on doit la réencoder
         String fileName = nomFichier;
         if (isVideoFile(fileName)) {
             String tempFilePath = basePath + dossierFichier + File.separator + fileName;
             storeFile(fileName, dossierFichier, base64);
             logger.info("InputPath: {}", tempFilePath );
-            // Réencoder la vidéo
-            String outputFilePath = basePath + dossierFichier + File.separator + "LIBEIL_" + fileName;
-            reencodeVideoToH264(tempFilePath, outputFilePath);
-            logger.info("OutputPath: {}", outputFilePath );
-            // Supprimer le fichier temporaire original après réencodage
-            new File(tempFilePath).delete();
+            if (!isH264Video(tempFilePath)){
+                String outputFilePath = basePath + dossierFichier + File.separator + "LIBEIL_" + fileName;
+                reencodeVideoToH264(tempFilePath, outputFilePath);
+                logger.info("OutputPath: {}", outputFilePath );
+                new File(tempFilePath).delete();
+            }else{
+                new File(tempFilePath).renameTo(new File(basePath + dossierFichier + File.separator + "LIBEIL_" + fileName));
+            }
         } else {
             // Si ce n'est pas une vidéo, on peut simplement stocker le fichier sans réencodage
             storeFile(fileName, dossierFichier, base64);
